refactor(publishing): move entry sorting to utilities
