+++ lib/get/get-full-source-space.js
@@ -1,7 +1,7 @@
 import Promise from 'bluebird'
 import Listr from 'listr'
 import verboseRenderer from 'listr-verbose-renderer'
-import sortEntries from './sort-entries'
+import sortEntries from '../utils/sort-entries'
 
 const MAX_ALLOWED_LIMIT = 1000
 let pageLimit = MAX_ALLOWED_LIMIT
