+++ lib/get/index.js
@@ -1,3 +1,2 @@
 export { default as getOutdatedDestinationContent } from './get-outdated-destination-content'
 export { default as getFullSourceSpace } from './get-full-source-space'
-export { default as sortEntries } from './sort-entries'
