+++ lib/transform/transform-space.js
@@ -2,7 +2,7 @@ import Promise from 'bluebird'
 import { omit, defaults } from 'lodash/object'
 
 import * as defaultTransformers from './transformers'
-import sortEntries from '../get/sort-entries'
+import sortEntries from '../utils/sort-entries'
 
 const spaceEntities = [
   'contentTypes', 'entries', 'assets', 'locales', 'webhooks'
