+++ lib/utils/index.js
@@ -1,2 +1,3 @@
 export { default as createClients } from './create-clients'
 export { default as errorBuffer } from './error-buffer'
+export { default as sortEntries } from './sort-entries'
