+++ lib/utils/sort-entries.js
